requirements.txt
argparse>=1.2.1
datadog
PyYAML
